Moved the actual aggregation of submissions into its own helper.
